requirements.txt
ipython==7.2.0
jira==2.0.0
python-dotenv==0.10.1
Flask==1.0.2
